[kern/kerx] Towards merge more
